make relative urls absolute to their source
if the rss reader tries to resolve a relative url like /episode/1, it
will try to resolve that relative to the css-rss website, which is
incorrect, as it should instead point to the source url.

the current check is quite primitive, as it only checks if the link
starts with a slash and therefore might miss relative links, that are
relative to the website with `./` for example, but i think that is rare
enough, that i will just leave it like this until someone complains.